Used session scoped fixture to setup and tear down tests (copy and delete xn files)
